+++ src/Hpack/Syntax/Defaults.hs
@@ -1,15 +1,19 @@
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RecordWildCards #-}
 module Hpack.Syntax.Defaults (
   Defaults(..)
+, DefaultsGithub(..)
+, DefaultsLocal(..)
 #ifdef TEST
 , isValidUser
 , isValidRepo
 #endif
 ) where
 
+import qualified Data.HashMap.Lazy as HM
 import           Data.List
 import qualified Data.Text as T
 import           System.FilePath.Posix (splitDirectories)
@@ -17,10 +21,10 @@ import           System.FilePath.Posix (splitDirectories)
 import           Data.Aeson.Config.FromValue
 import           Hpack.Syntax.Git
 
-data ParseDefaults = ParseDefaults {
-  parseDefaultsGithub :: Github
-, parseDefaultsRef :: Ref
-, parseDefaultsPath :: Maybe Path
+data ParseDefaultsGithub = ParseDefaultsGithub {
+  parseDefaultsGithubGithub :: Github
+, parseDefaultsGithubRef :: Ref
+, parseDefaultsGithubPath :: Maybe Path
 } deriving (Generic, FromValue)
 
 data Github = Github {
@@ -77,38 +81,57 @@ data Path = Path {unPath :: [FilePath]}
 
 instance FromValue Path where
   fromValue = withString parsePath
-    where
-      parsePath path
-        | '\\' `elem` path = fail ("rejecting '\\' in " ++ show path ++ ", please use '/' to separate path components")
-        | ':' `elem` path = fail ("rejecting ':' in " ++ show path)
-        | "/" `elem` p = fail ("rejecting absolute path " ++ show path)
-        | ".." `elem` p = fail ("rejecting \"..\" in " ++ show path)
-        | otherwise = return (Path p)
-        where
-          p = splitDirectories path
 
-data Defaults = Defaults {
+parsePath :: String -> Parser Path
+parsePath path
+  | '\\' `elem` path = fail ("rejecting '\\' in " ++ show path ++ ", please use '/' to separate path components")
+  | ':' `elem` path = fail ("rejecting ':' in " ++ show path)
+  | "/" `elem` p = fail ("rejecting absolute path " ++ show path)
+  | ".." `elem` p = fail ("rejecting \"..\" in " ++ show path)
+  | otherwise = return (Path p)
+  where
+    p = splitDirectories path
+
+data DefaultsGithub = DefaultsGithub {
   defaultsGithubUser :: String
 , defaultsGithubRepo :: String
-, defaultsRef :: String
-, defaultsPath :: [FilePath]
+, defaultsGithubRef :: String
+, defaultsGithubPath :: [FilePath]
 } deriving (Eq, Show)
 
-instance FromValue Defaults where
-  fromValue v = toDefaults <$> case v of
-    String s -> parseDefaultsFromString (T.unpack s)
-    Object _ -> fromValue v
-    _ -> typeMismatch "Object or String" v
-    where
-      toDefaults :: ParseDefaults -> Defaults
-      toDefaults ParseDefaults{..} = Defaults {
-          defaultsGithubUser = githubUser parseDefaultsGithub
-        , defaultsGithubRepo = githubRepo parseDefaultsGithub
-        , defaultsRef = unRef parseDefaultsRef
-        , defaultsPath = maybe [".hpack", "defaults.yaml"] unPath parseDefaultsPath
-        }
+toDefaultsGithub :: ParseDefaultsGithub -> DefaultsGithub
+toDefaultsGithub ParseDefaultsGithub{..} = DefaultsGithub {
+    defaultsGithubUser = githubUser parseDefaultsGithubGithub
+  , defaultsGithubRepo = githubRepo parseDefaultsGithubGithub
+  , defaultsGithubRef = unRef parseDefaultsGithubRef
+  , defaultsGithubPath = maybe [".hpack", "defaults.yaml"] unPath parseDefaultsGithubPath
+  }
 
-parseDefaultsFromString :: String -> Parser ParseDefaults
-parseDefaultsFromString xs = case break (== '@') xs of
-  (github, '@' : ref) -> ParseDefaults <$> parseGithub github <*> parseRef ref <*> pure Nothing
+parseDefaultsGithubFromString :: String -> Parser ParseDefaultsGithub
+parseDefaultsGithubFromString xs = case break (== '@') xs of
+  (github, '@' : ref) -> ParseDefaultsGithub <$> parseGithub github <*> parseRef ref <*> pure Nothing
   _ -> fail ("missing Git reference for " ++ show xs ++ ", the expected format is user/repo@ref")
+
+data ParseDefaultsLocal = ParseDefaultsLocal {
+  parseDefaultsLocalLocal :: FilePath
+} deriving (Generic, FromValue)
+
+data DefaultsLocal = DefaultsLocal {
+  defaultsLocal :: String
+} deriving (Eq, Show)
+
+toDefaultsLocal :: ParseDefaultsLocal -> DefaultsLocal
+toDefaultsLocal = DefaultsLocal . parseDefaultsLocalLocal
+
+data Defaults = DefaultsLocal_ DefaultsLocal | DefaultsGithub_ DefaultsGithub
+  deriving (Eq, Show)
+
+instance FromValue Defaults where
+  fromValue v = case v of
+    String s -> DefaultsGithub_ . toDefaultsGithub <$> parseDefaultsGithubFromString (T.unpack s)
+    Object o -> case (HM.member "local" o, HM.member "github" o) of
+      (False, False) -> fail "must have exactly one of these keys: \"local\" or \"github\""
+      (True, True) -> fail "both \"github\" and \"local\" are present; please use one or the other."
+      (True, False) -> DefaultsLocal_ . toDefaultsLocal <$> fromValue v
+      (False, True) -> DefaultsGithub_ . toDefaultsGithub <$> fromValue v
+    _ -> typeMismatch "Object or String" v