+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - Remove Git conflict markers before checking the hash of any existing
     `.cabal` files (equivalent to `git checkout --ours`).  This allows you to
     regenerate the `.cabal` on conflicts when rebasing without passing `-f`.
+  - Allow local files to be used as defaults (#248)
 
 ## Changes in 0.25.0
   - Keep non-existing literal files on glob expansion (see #101)